requirements.txt
tensorflow-macos
numpy
Pillow
scikit-learn
tk
python-docx
